requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
python-dotenv>=1.0
APScheduler>=3.10
requests>=2.32
beautifulsoup4>=4.12
bleach>=6.1
pydantic>=2.7
